enhance: 拆分动画到组件
