+++ frontend_vue/src/components/views/MainMenu.vue
@@ -11,38 +11,10 @@
     <div class="video-background" ref="bgRef"></div>
 
     <!-- 流星层（SVG动画） -->
-    <div v-if="meteorsEnabled" class="meteor-wrapper">
-      <svg id="meteor-svg" ref="svgRef">
-        <defs>
-          <linearGradient
-            id="meteor-grad-fixed"
-            gradientUnits="objectBoundingBox"
-            x1="1"
-            y1="0"
-            x2="0"
-            y2="0"
-          >
-            <stop offset="0%" stop-color="rgba(255,255,255,1)" />
-            <stop offset="40%" stop-color="rgba(255,255,255,1)" />
-            <stop offset="70%" stop-color="rgba(220,240,255,0.8)" />
-            <stop offset="100%" stop-color="rgba(180,220,255,0)" />
-          </linearGradient>
-        </defs>
-        <!-- Vue 响应式渲染流星路径 -->
-        <path
-          v-for="meteor in activeMeteors"
-          :key="meteor.id"
-          :d="meteor.d"
-          class="meteor-path"
-          :style="{ animationDuration: `${meteor.duration}s` }"
-        />
-      </svg>
-    </div>
+    <MeteorAnimation :meteors-enabled="meteorsEnabled" />
 
     <!-- 星星粒子层（位于背景和人物之间） -->
-    <div v-if="starsEnabled" class="stars-layer" ref="starsLayerRef">
-      <canvas id="stars-canvas" ref="canvasRef"></canvas>
-    </div>
+    <StarAnimation :stars-enabled="starsEnabled" :stars-layer-ref="starsLayerRef" />
 
     <!-- 人物图层（位于星星之上，菜单之下） -->
     <img class="character-image" ref="charRef" src="../../assets/images/alona.png" alt="人物" />
@@ -106,6 +78,8 @@ import { useSettingsStore } from '../../stores/modules/settings'
 import ScriptModeOptions from './menu/ScriptModeOptions.vue'
 import { getScriptList, type ScriptSummary } from '@/api/services/script-info'
 import { saveContinue } from '@/api/services/save'
+import MeteorAnimation from '../game/standard/animations/MeteorAnimation.vue'
+import StarAnimation from '../game/standard/animations/StarAnimation.vue'
 
 const router = useRouter()
 const uiStore = useUIStore()
@@ -124,7 +98,6 @@ const containerRef = ref<HTMLElement | null>(null)
 const bgRef = ref<HTMLElement | null>(null)
 const charRef = ref<HTMLElement | null>(null)
 const starsLayerRef = ref<HTMLElement | null>(null)
-const canvasRef = ref<HTMLCanvasElement | null>(null)
 
 const Save = Settings
 
@@ -275,339 +248,6 @@ function handleMouseLeave() {
   startParallaxIfNeeded()
 }
 
-/* ================== 星星粒子系统 (优化版) ================== */
-interface Star {
-  x: number
-  y: number
-  size: number
-  baseOpacity: number
-  rotation: number
-  type: 'star' | 'circle'
-}
-
-const STARS_COUNT = 80
-const FLICKER_SPEED = 0.003
-const starsPositions = shallowRef<Star[]>([])
-let starsFrameId: number | null = null
-let starsCtx: CanvasRenderingContext2D | null = null
-
-// 缓存预渲染的星星图像
-let starImageCache: Map<number, HTMLCanvasElement> | null = null
-let circleImageCache: Map<number, HTMLCanvasElement> | null = null
-
-/**
- * 创建带发光效果的星星形状到离屏 canvas
- * 预渲染一次，后续使用 drawImage 快速绘制
- */
-function createStarImage(size: number): HTMLCanvasElement {
-  const padding = size * 2 // 为发光效果预留空间
-  const canvas = document.createElement('canvas')
-  const actualSize = size * 1.5 + padding * 2
-  canvas.width = actualSize
-  canvas.height = actualSize
-  const ctx = canvas.getContext('2d')!
-  const centerX = actualSize / 2
-  const centerY = actualSize / 2
-
-  // 发光效果
-  ctx.shadowColor = 'rgba(255, 255, 200, 0.8)'
-  ctx.shadowBlur = 20
-
-  // 渐变填充
-  const gradient = ctx.createRadialGradient(centerX, centerY, 0, centerX, centerY, size * 1.5)
-  gradient.addColorStop(0, '#ffffff')
-  gradient.addColorStop(0.6, 'rgba(255, 240, 200, 0.9)')
-  gradient.addColorStop(1, 'rgba(255, 200, 100, 0)')
-  ctx.fillStyle = gradient
-
-  // 绘制星形路径
-  ctx.beginPath()
-  for (let i = 0; i < 8; i++) {
-    const angle = (i * Math.PI) / 4
-    const radius = i % 2 === 0 ? size : size * 0.4
-    const dx = centerX + Math.cos(angle) * radius
-    const dy = centerY + Math.sin(angle) * radius
-    if (i === 0) ctx.moveTo(dx, dy)
-    else ctx.lineTo(dx, dy)
-  }
-  ctx.closePath()
-  ctx.fill()
-
-  return canvas
-}
-
-/**
- * 创建带发光效果的圆形到离屏 canvas
- */
-function createCircleImage(size: number): HTMLCanvasElement {
-  const padding = size * 2
-  const canvas = document.createElement('canvas')
-  const actualSize = size * 1.5 + padding * 2
-  canvas.width = actualSize
-  canvas.height = actualSize
-  const ctx = canvas.getContext('2d')!
-  const centerX = actualSize / 2
-  const centerY = actualSize / 2
-
-  // 发光效果
-  ctx.shadowColor = 'rgba(255, 255, 200, 0.8)'
-  ctx.shadowBlur = 20
-
-  // 渐变填充
-  const gradient = ctx.createRadialGradient(centerX, centerY, 0, centerX, centerY, size * 1.5)
-  gradient.addColorStop(0, '#ffffff')
-  gradient.addColorStop(0.6, 'rgba(255, 240, 200, 0.9)')
-  gradient.addColorStop(1, 'rgba(255, 200, 100, 0)')
-  ctx.fillStyle = gradient
-
-  ctx.beginPath()
-  ctx.arc(centerX, centerY, size, 0, Math.PI * 2)
-  ctx.fill()
-
-  return canvas
-}
-
-/**
- * 获取或创建缓存的星星图像
- * 使用 Map 缓存不同尺寸的预渲染图像
- */
-function getStarImage(size: number): HTMLCanvasElement {
-  if (!starImageCache) {
-    starImageCache = new Map()
-  }
-  const roundedSize = Math.round(size)
-  let image = starImageCache.get(roundedSize)
-  if (!image) {
-    image = createStarImage(roundedSize)
-    starImageCache.set(roundedSize, image)
-  }
-  return image
-}
-
-function getCircleImage(size: number): HTMLCanvasElement {
-  if (!circleImageCache) {
-    circleImageCache = new Map()
-  }
-  const roundedSize = Math.round(size)
-  let image = circleImageCache.get(roundedSize)
-  if (!image) {
-    image = createCircleImage(roundedSize)
-    circleImageCache.set(roundedSize, image)
-  }
-  return image
-}
-
-/**
- * 清理图像缓存
- */
-function clearImageCaches() {
-  starImageCache?.clear()
-  circleImageCache?.clear()
-  starImageCache = null
-  circleImageCache = null
-}
-
-function generateStars() {
-  if (!canvasRef.value) return
-  const w = window.innerWidth
-  const h = window.innerHeight
-  canvasRef.value.width = w
-  canvasRef.value.height = h
-  starsCtx = canvasRef.value.getContext('2d')
-  if (!starsCtx) return
-
-  // 预生成所有可能用到的尺寸的图像缓存
-  // 星星尺寸范围是 2-7 (Math.random() * 5 + 2)
-  // 预生成常用尺寸以减少运行时开销
-  for (let size = 2; size <= 7; size++) {
-    getStarImage(size)
-    getCircleImage(size)
-  }
-
-  const tempStars: Star[] = []
-  for (let i = 0; i < STARS_COUNT; i++) {
-    tempStars.push({
-      x: Math.random() * w,
-      y: Math.random() * (h * 0.33),
-      size: Math.random() * 5 + 2,
-      baseOpacity: Math.random() * 0.5 + 0.5,
-      rotation: Math.random() * Math.PI * 2,
-      type: Math.random() > 0.2 ? 'star' : 'circle',
-    })
-  }
-  starsPositions.value = tempStars
-}
-
-/**
- * 优化版渲染函数
- * - 使用预缓存的离屏 canvas 图像
- * - 避免每帧创建渐变和设置阴影
- * - 缓存 stars 数组避免响应式开销
- */
-function renderStars() {
-  if (!starsCtx || !canvasRef.value) return
-  const w = canvasRef.value.width
-  const h = canvasRef.value.height
-
-  // 关键修复：重置全局透明度再 clearRect，避免清除不干净导致的问题
-  starsCtx.globalAlpha = 1.0
-  starsCtx.clearRect(0, 0, w, h)
-
-  const now = Date.now()
-
-  const stars = starsPositions.value
-
-  for (let i = 0; i < stars.length; i++) {
-    const pos = stars[i]
-    if (!pos) continue
-
-    const flicker = 0.6 + 0.4 * Math.sin(now * FLICKER_SPEED + pos.x)
-    const opacity = Math.min(pos.baseOpacity * flicker, 1.0)
-    starsCtx!.globalAlpha = opacity
-
-    // 获取预缓存的图像
-    const image = pos.type === 'star' ? getStarImage(pos.size) : getCircleImage(pos.size)
-    const imgSize = image.width
-    const halfSize = imgSize / 2
-
-    // 使用 drawImage 快速绘制，避免每帧创建渐变和路径
-    if (pos.type === 'star') {
-      // 星星需要旋转
-      starsCtx.save()
-      starsCtx.translate(pos.x, pos.y)
-      starsCtx.rotate(pos.rotation)
-      starsCtx.drawImage(image, -halfSize, -halfSize)
-      starsCtx.restore()
-    } else {
-      // 圆形无需旋转，直接绘制
-      starsCtx.drawImage(image, pos.x - halfSize, pos.y - halfSize)
-    }
-  }
-}
-
-function flickerAnimation() {
-  renderStars()
-  starsFrameId = requestAnimationFrame(flickerAnimation)
-}
-
-function handleResize() {
-  stopStars()
-  startStars()
-}
-
-/* ================== 流星雨系统 ================== */
-interface MeteorData {
-  id: number
-  d: string
-  duration: number
-  startY: number
-}
-const activeMeteors = ref<MeteorData[]>([])
-let meteorIdCounter = 0
-let meteorIntervalId: ReturnType<typeof setInterval> | null = null
-const METEOR_CONFIG = {
-  MAX_COUNT: 3,
-  GENERATE_INTERVAL: 1000,
-  MIN_DISTANCE: 800,
-  DURATION_MIN: 9,
-  DURATION_MAX: 14,
-  START_X_MIN: 800,
-  START_X_MAX: 1200,
-  START_Y_RANGE: 0.25,
-  END_X_MIN: -200,
-  END_X_MAX: -400,
-  END_Y_OFFSET: 1,
-  CONTROL_X_RATIO_MIN: 0.5,
-  CONTROL_X_RATIO_MAX: 0.9,
-  CONTROL_Y_OFFSET: -0.1,
-}
-
-function getValidStartY() {
-  const h = window.innerHeight
-  const maxAttempts = 120
-  let attempts = 0
-  while (attempts < maxAttempts) {
-    const startY = -Math.random() * (h * METEOR_CONFIG.START_Y_RANGE)
-    const tooClose = activeMeteors.value.some(
-      (m) => Math.abs(startY - m.startY) < METEOR_CONFIG.MIN_DISTANCE,
-    )
-    if (!tooClose) return startY
-    attempts++
-  }
-  return -Math.random() * (h * METEOR_CONFIG.START_Y_RANGE)
-}
-
-function createMeteor() {
-  const w = window.innerWidth
-  const h = window.innerHeight
-  const startY = getValidStartY()
-  const startX =
-    w +
-    METEOR_CONFIG.START_X_MIN +
-    Math.random() * (METEOR_CONFIG.START_X_MAX - METEOR_CONFIG.START_X_MIN)
-  const endX =
-    METEOR_CONFIG.END_X_MIN + Math.random() * (METEOR_CONFIG.END_X_MAX - METEOR_CONFIG.END_X_MIN)
-  const endY = startY + h * METEOR_CONFIG.END_Y_OFFSET
-  const controlX =
-    w *
-    (METEOR_CONFIG.CONTROL_X_RATIO_MIN +
-      Math.random() * (METEOR_CONFIG.CONTROL_X_RATIO_MAX - METEOR_CONFIG.CONTROL_X_RATIO_MIN))
-  const controlY = startY + h * METEOR_CONFIG.CONTROL_Y_OFFSET
-  const duration =
-    METEOR_CONFIG.DURATION_MIN +
-    Math.random() * (METEOR_CONFIG.DURATION_MAX - METEOR_CONFIG.DURATION_MIN)
-  const d = `M ${startX} ${startY} Q ${controlX} ${controlY} ${endX} ${endY}`
-  const id = meteorIdCounter++
-
-  activeMeteors.value.push({ id, d, duration, startY })
-  setTimeout(() => {
-    activeMeteors.value = activeMeteors.value.filter((m) => m.id !== id)
-  }, duration * 1000)
-}
-
-function updateMeteorShower() {
-  if (activeMeteors.value.length < METEOR_CONFIG.MAX_COUNT) createMeteor()
-}
-
-function startMeteorShower() {
-  if (meteorIntervalId) clearInterval(meteorIntervalId)
-  for (let i = 0; i < 3; i++) {
-    setTimeout(() => {
-      if (activeMeteors.value.length < METEOR_CONFIG.MAX_COUNT) createMeteor()
-    }, i * 300)
-  }
-  meteorIntervalId = setInterval(updateMeteorShower, METEOR_CONFIG.GENERATE_INTERVAL)
-}
-
-function startStars() {
-  if (!canvasRef.value) return
-  generateStars()
-  flickerAnimation()
-  window.removeEventListener('resize', handleResize)
-  window.addEventListener('resize', handleResize)
-}
-
-function stopStars() {
-  window.removeEventListener('resize', handleResize)
-  if (starsFrameId) {
-    cancelAnimationFrame(starsFrameId)
-    starsFrameId = null
-  }
-  if (starsCtx && canvasRef.value) {
-    starsCtx.clearRect(0, 0, canvasRef.value.width, canvasRef.value.height)
-  }
-  // 清理图像缓存
-  clearImageCaches()
-}
-
-function stopMeteorShower() {
-  if (meteorIntervalId) {
-    clearInterval(meteorIntervalId)
-    meteorIntervalId = null
-  }
-  activeMeteors.value = []
-}
-
 /* ================== 生命周期Hooks ================== */
 
 // 抽取接口请求逻辑，不阻塞动画初始化
@@ -646,36 +286,9 @@ onMounted(() => {
   }
 
   initializeMenu()
-
-  watch(
-    starsEnabled,
-    async (enabled) => {
-      if (enabled) {
-        await nextTick()
-        startStars()
-      } else {
-        stopStars()
-      }
-    },
-    { immediate: true },
-  )
-
-  watch(
-    meteorsEnabled,
-    (enabled) => {
-      if (enabled) {
-        startMeteorShower()
-      } else {
-        stopMeteorShower()
-      }
-    },
-    { immediate: true },
-  )
 })
 
 onUnmounted(() => {
-  stopStars()
-  stopMeteorShower()
   if (parallaxRafId) cancelAnimationFrame(parallaxRafId)
 })
 </script>
@@ -787,65 +400,5 @@ onUnmounted(() => {
   pointer-events: none;
   /* 移除 transition */
   will-change: transform;
-}
-
-/* ========== 星星粒子层 ========== */
-.stars-layer {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 1;
-  pointer-events: none;
-  overflow: hidden;
-  /* 移除 transition */
-  will-change: transform;
-}
-
-#stars-canvas {
-  width: 100%;
-  height: 100%;
-  display: block;
-}
-
-/* ========== 流星层 ========== */
-.meteor-wrapper {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 2;
-  pointer-events: none;
-  overflow: hidden;
-}
-
-#meteor-svg {
-  width: 100%;
-  height: 100%;
-}
-
-.meteor-path {
-  fill: none;
-  stroke: url(#meteor-grad-fixed);
-  stroke-width: 5;
-  stroke-linecap: round;
-  filter: drop-shadow(0 0 20px rgba(255, 255, 255, 0.95))
-    drop-shadow(0 0 12px rgba(180, 220, 255, 0.9));
-  stroke-dasharray: 600, 3000;
-  stroke-dashoffset: 3000;
-  animation-name: meteor-move;
-  animation-timing-function: linear;
-  animation-fill-mode: forwards;
-}
-
-@keyframes meteor-move {
-  0% {
-    stroke-dashoffset: 3000;
-  }
-  100% {
-    stroke-dashoffset: -3000;
-  }
 }
 </style>
